+++ cli/js/testing.ts
@@ -14,6 +14,12 @@ const GREEN_OK = green("ok");
 const YELLOW_IGNORED = yellow("ignored");
 const disabledConsole = new Console((): void => {});
 
+function delay(n: number): Promise<void> {
+  return new Promise((resolve: () => void, _) => {
+    setTimeout(resolve, n);
+  });
+}
+
 function formatDuration(time = 0): string {
   const timeStr = `(${time}ms)`;
   return gray(italic(timeStr));
@@ -28,6 +34,10 @@ function assertOps(fn: () => void | Promise<void>): () => void | Promise<void> {
   return async function asyncOpSanitizer(): Promise<void> {
     const pre = metrics();
     await fn();
+    // Defer until next event loop turn - that way timeouts and intervals
+    // cleared can actually be removed from resource table, otherwise
+    // false positives may occur (https://github.com/denoland/deno/issues/4591)
+    await delay(0);
     const post = metrics();
     // We're checking diff because one might spawn HTTP server in the background
     // that will be a pending async op before test starts.