+++ cli/lib.rs
@@ -61,11 +61,16 @@ pub mod version;
 mod web_worker;
 pub mod worker;
 
+pub use dprint_plugin_typescript::swc_common;
+pub use dprint_plugin_typescript::swc_ecma_ast;
+pub use dprint_plugin_typescript::swc_ecma_parser;
+
 use crate::compilers::TargetLib;
 use crate::file_fetcher::SourceFile;
 use crate::global_state::GlobalState;
 use crate::msg::MediaType;
 use crate::ops::io::get_stdio;
+use crate::state::DebugType;
 use crate::state::State;
 use crate::worker::MainWorker;
 use deno_core::v8_set_flags;
@@ -128,7 +133,7 @@ fn create_main_worker(
   global_state: GlobalState,
   main_module: ModuleSpecifier,
 ) -> Result<MainWorker, ErrBox> {
-  let state = State::new(global_state, None, main_module)?;
+  let state = State::new(global_state, None, main_module, DebugType::Main)?;
 
   {
     let mut s = state.borrow_mut();