+++ std/node/fs.ts
@@ -8,6 +8,7 @@ import { close, closeSync } from "./_fs/_fs_close.ts";
 import * as constants from "./_fs/_fs_constants.ts";
 import { readFile, readFileSync } from "./_fs/_fs_readFile.ts";
 import { readlink, readlinkSync } from "./_fs/_fs_readlink.ts";
+import { exists, existsSync } from "./_fs/_fs_exists.ts";
 
 export {
   access,
@@ -25,4 +26,6 @@ export {
   readFileSync,
   readlink,
   readlinkSync,
+  exists,
+  existsSync,
 };