+++ Releases.md
@@ -6,6 +6,39 @@ https://github.com/denoland/deno/releases
 We also have one-line install commands at
 https://github.com/denoland/deno_install
 
+### v0.39.0 / 2020.04.03
+
+- BREAKING CHANGE: Move encode, decode helpers to /std/encoding/utf8.ts, delete
+  /std/strings/ (#4565)
+- BREAKING CHANGE: Remove /std/media_types (#4594)
+- BREAKING CHANGE: Remove old release files (#4545)
+- BREAKING CHANGE: Remove std/strings/pad.ts because String.prototype.padStart
+  exists (#4564)
+- feat: Add common to std/path (#4527)
+- feat: Added colors to doc output (#4518)
+- feat: Expose global state publicly (#4572)
+- feat: Make inspector more robust, add --inspect-brk support (#4552)
+- feat: Publish deno types on release (#4583)
+- feat: Support dynamic import in bundles. (#4561)
+- feat: deno test --filter (#4570)
+- feat: improve console.log serialization (#4524, #4472)
+- fix(#4550): setCookie should append cookies (#4558)
+- fix(#4554): use --inspect in repl & eval (#4562)
+- fix(deno doc): handle 'declare' (#4573)
+- fix(deno doc): parse super-class names (#4595)
+- fix(deno doc): parse the "implements" clause of a class def (#4604)
+- fix(file_server): serve appropriate content-type header (#4555)
+- fix(inspector): proper error message on port collision (#4514)
+- fix: Add check to fail the benchmark test on server error (#4519)
+- fix: Properly handle invalid utf8 in paths (#4609)
+- fix: async ops sanitizer false positives in timers (#4602)
+- fix: invalid blob type (#4536)
+- fix: make Worker.poll private (#4603)
+- fix: remove `Send` trait requirement from the `Resource` trait (#4585)
+- refactor(testing): Reduce testing interfaces (#4451)
+- upgrade: dprint to 0.9.10 (#4601)
+- upgrade: rusty_v8 v0.3.10 (#4576)
+
 ### v0.38.0 / 2020.03.28
 
 - feat: Add "deno doc" subcommand (#4500)